+++ src/main/java/io/github/mapepire_ibmi/MapepireResultSet.java
@@ -343,14 +343,17 @@ public class MapepireResultSet implements ResultSet {
 
     @Override
     public SQLWarning getWarnings() throws SQLException {
-        // TODO Auto-generated method stub
-        throw new SQLFeatureNotSupportedException("Unimplemented method 'getWarnings'");
+        if (this.closed) {
+            throw new SQLException("ResultSet is closed");
+        }
+        return null;
     }
 
     @Override
     public void clearWarnings() throws SQLException {
-        // TODO Auto-generated method stub
-        throw new SQLFeatureNotSupportedException("Unimplemented method 'clearWarnings'");
+        if (this.closed) {
+            throw new SQLException("ResultSet is closed");
+        }
     }
 
     @Override